+++ src/com/oltpbenchmark/benchmarks/tpcc/procedures/NewOrder.java
@@ -36,16 +36,16 @@ public class NewOrder extends TPCCProcedure {
 
     public final SQLStmt stmtGetCustSQL = new SQLStmt(
     		"SELECT C_DISCOUNT, C_LAST, C_CREDIT" +
-	        "  FROM " + TPCCConstants.TABLENAME_CUSTOMER + 
-	        " WHERE C_W_ID = ? " + 
+	        "  FROM " + TPCCConstants.TABLENAME_CUSTOMER +
+	        " WHERE C_W_ID = ? " +
 	        "   AND C_D_ID = ? " +
 	        "   AND C_ID = ?");
 
     public final SQLStmt stmtGetWhseSQL = new SQLStmt(
-    		"SELECT W_TAX " + 
-		    "  FROM " + TPCCConstants.TABLENAME_WAREHOUSE + 
+    		"SELECT W_TAX " +
+		    "  FROM " + TPCCConstants.TABLENAME_WAREHOUSE +
 		    " WHERE W_ID = ?");
-    
+
     public final SQLStmt stmtGetDistSQL = new SQLStmt(
     		"SELECT D_NEXT_O_ID, D_TAX " +
 	        "  FROM " + TPCCConstants.TABLENAME_DISTRICT +
@@ -57,39 +57,39 @@ public class NewOrder extends TPCCProcedure {
             " VALUES ( ?, ?, ?)");
 
 	public final SQLStmt  stmtUpdateDistSQL = new SQLStmt(
-	        "UPDATE " + TPCCConstants.TABLENAME_DISTRICT + 
+	        "UPDATE " + TPCCConstants.TABLENAME_DISTRICT +
 	        "   SET D_NEXT_O_ID = D_NEXT_O_ID + 1 " +
             " WHERE D_W_ID = ? " +
 	        "   AND D_ID = ?");
 
 	public final SQLStmt  stmtInsertOOrderSQL = new SQLStmt(
-	        "INSERT INTO " + TPCCConstants.TABLENAME_OPENORDER + 
-	        " (O_ID, O_D_ID, O_W_ID, O_C_ID, O_ENTRY_D, O_OL_CNT, O_ALL_LOCAL)" + 
+	        "INSERT INTO " + TPCCConstants.TABLENAME_OPENORDER +
+	        " (O_ID, O_D_ID, O_W_ID, O_C_ID, O_ENTRY_D, O_OL_CNT, O_ALL_LOCAL)" +
             " VALUES (?, ?, ?, ?, ?, ?, ?)");
 
 	public final SQLStmt  stmtGetItemSQL = new SQLStmt(
 	        "SELECT I_PRICE, I_NAME , I_DATA " +
-            "  FROM " + TPCCConstants.TABLENAME_ITEM + 
+            "  FROM " + TPCCConstants.TABLENAME_ITEM +
             " WHERE I_ID = ?");
 
 	public final SQLStmt  stmtGetStockSQL = new SQLStmt(
 	        "SELECT S_QUANTITY, S_DATA, S_DIST_01, S_DIST_02, S_DIST_03, S_DIST_04, S_DIST_05, " +
             "       S_DIST_06, S_DIST_07, S_DIST_08, S_DIST_09, S_DIST_10" +
-            "  FROM " + TPCCConstants.TABLENAME_STOCK + 
+            "  FROM " + TPCCConstants.TABLENAME_STOCK +
             " WHERE S_I_ID = ? " +
             "   AND S_W_ID = ? FOR UPDATE");
 
 	public final SQLStmt  stmtUpdateStockSQL = new SQLStmt(
-	        "UPDATE " + TPCCConstants.TABLENAME_STOCK + 
+	        "UPDATE " + TPCCConstants.TABLENAME_STOCK +
 	        "   SET S_QUANTITY = ? , " +
-            "       S_YTD = S_YTD + ?, " + 
+            "       S_YTD = S_YTD + ?, " +
 	        "       S_ORDER_CNT = S_ORDER_CNT + 1, " +
             "       S_REMOTE_CNT = S_REMOTE_CNT + ? " +
 	        " WHERE S_I_ID = ? " +
             "   AND S_W_ID = ?");
 
 	public final SQLStmt  stmtInsertOrderLineSQL = new SQLStmt(
-	        "INSERT INTO " + TPCCConstants.TABLENAME_ORDERLINE + 
+	        "INSERT INTO " + TPCCConstants.TABLENAME_ORDERLINE +
 	        " (OL_O_ID, OL_D_ID, OL_W_ID, OL_NUMBER, OL_I_ID, OL_SUPPLY_W_ID, OL_QUANTITY, OL_AMOUNT, OL_DIST_INFO) " +
             " VALUES (?,?,?,?,?,?,?,?,?)");
 
@@ -130,7 +130,7 @@ public class NewOrder extends TPCCProcedure {
 		int districtID = TPCCUtil.randomNumber(terminalDistrictLowerID,terminalDistrictUpperID, gen);
 		int customerID = TPCCUtil.getCustomerID(gen);
 
-		int numItems = (int) TPCCUtil.randomNumber(5, 15, gen);
+		int numItems = TPCCUtil.randomNumber(5, 15, gen);
 		int[] itemIDs = new int[numItems];
 		int[] supplierWarehouseIDs = new int[numItems];
 		int[] orderQuantities = new int[numItems];
@@ -182,7 +182,7 @@ public class NewOrder extends TPCCProcedure {
 		int ol_supply_w_id, ol_i_id, ol_quantity;
 		int s_remote_cnt_increment;
 		float ol_amount, total_amount = 0;
-		
+
 		try {
 			stmtGetCust.setInt(1, w_id);
 			stmtGetCust.setInt(2, d_id);
@@ -249,7 +249,7 @@ public class NewOrder extends TPCCProcedure {
 			/*TODO: add error checking */
 
 
-			/* woonhak, [[change order				 
+			/* woonhak, [[change order
 			stmtInsertOOrder.setInt(1, o_id);
 			stmtInsertOOrder.setInt(2, d_id);
 			stmtInsertOOrder.setInt(3, w_id);
@@ -342,7 +342,7 @@ public class NewOrder extends TPCCProcedure {
 					brandGeneric[ol_number - 1] = 'G';
 				}
 
-				switch ((int) d_id) {
+				switch (d_id) {
 				case 1:
 					ol_dist_info = s_dist_01;
 					break;
